requirements.txt
blessed==1.20.0
botocore==1.35.13
chalice==1.31.2
click==8.1.7
inquirer==2.10.1
jmespath==1.0.1
Pillow==8.3.1
python-dateutil==2.9.0.post0
python-editor==1.0.4
PyYAML==6.0.2
readchar==4.2.0
six==1.16.0
urllib3==1.26.20
wcwidth==0.2.13

